# linux/aarch64/frontend/docs/openapi.yml
openapi: "3.0.3"
info:
  title: OmniHand API
  version: "1.1.0"
  license:
    name: Mulan PSL v2
    url: https://license.coscl.org.cn/MulanPSL2
  description: |
    OmniHand SDK REST API — 手部实例管理与远程方法调用。

    支持产品型号:
    - `omnihand_2025` (O10, 10 DOF)
    - `omnihand_pro_2025` (O12, 12 DOF)
    - `omnihand_dex_umi` (10 DOF, UMI 协议)
    - `omnihand_3_lite` (O4, 4 DOF)
    - `omnihand_3_ultra_m` (O20, 20 DOF)
  contact:
    name: OmniHand SDK Team
    email: junxiaowu@agibot.com
servers:
  - url: http://localhost:8000/v1
    description: 本地开发服务器
security: []
paths:
  /hands:
    post:
      tags: [Hands]
      summary: 创建 OmniHand 手部实例
      operationId: create_hand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandCreateRequest"
      responses:
        "200":
          description: 手部实例已存在（幂等创建）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandCreateResponse"
        "201":
          description: 成功创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandCreateResponse"
        "400":
          description: 无效请求参数
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Hands]
      summary: 获取所有已创建的手部实例
      operationId: get_hands
      responses:
        "200":
          description: 手部实例列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  hands:
                    type: array
                    items:
                      $ref: "#/components/schemas/HandSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  schemas:
...
    HandCreateRequest:
      type: object
      required: [hand_type, conn_method, conn_config]
      properties:
        hand_type:
          type: string
          description: 手部型号
          enum:
            - omnihand_2025
            - omnihand_pro_2025
            - omnihand_dex_umi
            - omnihand_3_lite
            - omnihand_3_ultra_m
        hand_side:
          type: string
          description: 手部侧别，默认为左手
          default: left
          enum: [left, right]
        conn_method:
          type: string
          description: 连接方式
          enum:
            - zlgcan
            - hcan
            - socketcan
            - rs485
            - usb
            - zlgcan_tcp
        conn_config:
          type: object
          description: |
            连接参数，取决于 conn_method:
            - **zlgcan**: `{hand_type, hand_device_id, canfd_device_id, canfd_channel_id}` 或 `{hand_type, hand_device_id, usbcanfd_serial_number, canfd_channel_id}`
            - **hcan**: `{hand_type, hand_device_id, canfd_device_id, canfd_channel_id}` 或 `{hand_type, hand_device_id, hcan_serial_number, canfd_channel_id}`
            - **socketcan**: `{hand_type, hand_device_id, can_interface}`
            - **rs485**: `{hand_type, hand_device_id, uart_port, baudrate}`
            - **usb**: `{hand_type, hand_device_id, uart_port, baudrate}`
            - **zlgcan_tcp**: `{hand_type, hand_device_id, host, port, canfd_channel_id}`

    # ========== 响应 ==========
    HandCreateResponse:
      type: object
      required: [hand_id, already_created]
      properties:
        hand_id:
          type: integer
          description: 后端注册的唯一手部标识
        already_created:
          type: boolean
          description: 该 hand_id 是否已存在（幂等创建）
    HandSummary:
      type: object
      properties:
        hand_id:
          type: integer
        hand_type:
          type: string
        conn_method:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
    BadRequest:
      description: 请求参数无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...